==> Cryptokitties/1_getting kitties_prices.ipynb <==
import requests
import csv
import json
import pandas as pd
import numpy as np
from tqdm import tqdm
%matplotlib inline

# Getting kitties prices

import time
res=[]
i=0
for off in tqdm(range(0,164990,100)):
    time.sleep(2)
    base_url = 'https://kittysales.herokuapp.com/data?offset=%d&count=100&sortBy=timestamp&direction=desc' % (off)
    response = requests.get(base_url)
    results = response.json()
    res.append(results)
    i+=1

i

#первый блок первая транзакция
res[0]['sales'][0]

#нужные колонки
cat_id = res[0]['sales'][0]['returnValues']['0'] 
soldPriceUSD = res[0]['sales'][0]['soldPrice']
timeStamp = res[1427]['sales'][99]['blocktimeStamp']

### Write necessary columns to file

f = csv.writer(open("cat_price160k.csv", "w"))
# header
f.writerow(["cat_id", "soldPrice", "timeStamp"])

for i in range(1650):
    for j in range(100):
        f.writerow([res[i]['sales'][j]['returnValues']['0'],
              res[i]['sales'][j]['soldPrice'],
              res[i]['sales'][j]['blocktimeStamp']])


price_df = pd.read_csv('cat_price160k.csv')

price_df.shape

price_df[price_df['cat_id']==1189]

### Check for dublicates

pd.concat(g for _, g in price_df.groupby("cat_id") if len(g) > 1)

### Leave only last price

clear_price=price_df.sort_values('timeStamp').drop_duplicates(subset='cat_id', keep='last')

#check
clear_price[clear_price['cat_id']==2]

clear_price.shape

#save to csv
clear_price.to_csv('clear_price160k.csv')

# NEXT prices

import time
res2=[]
i=0
for off in tqdm(range(164990,192754,100)):
    time.sleep(2)
    base_url = 'https://kittysales.herokuapp.com/data?offset=%d&count=100&sortBy=timestamp&direction=desc' % (off)
    response = requests.get(base_url)
    results = response.json()
    res2.append(results)
    i+=1

f = csv.writer(open("cat_price160_190k.csv", "w"))
# header
f.writerow(["cat_id", "soldPrice", "timeStamp"])

for i in range(111):
    for j in range(100):
        f.writerow([res2[i]['sales'][j]['returnValues']['0'],
              res2[i]['sales'][j]['soldPrice'],
              res2[i]['sales'][j]['blocktimeStamp']])


price_df = pd.read_csv('cat_price160k.csv')
price_df2 = pd.read_csv('cat_price160_190k.csv')

price_df2.shape

result = pd.concat([price_df,price_df2])

result.shape

clear_price=result.sort_values('timeStamp').drop_duplicates(subset='cat_id', keep='last')

clear_price.tail()

import datetime
df=clear_price
df['dateSold'] = df['timeStamp'].apply(datetime.datetime.fromtimestamp)
df['date']=df['dateSold'].apply(lambda x : x.date())

df = df.sort_values('date')

df.tail()

result.to_csv('unclear_price190k.csv', index=False)

clear_price.to_csv('clear_price190k.csv', index=False)
