==> Cryptokitties/2_getting fancy.ipynb <==
from urllib.request import Request, urlopen
import time
import csv
import json

import re

# Chicken lst

chick_lst=[]

for i in range(1,171):
    time.sleep(4)
    req = Request('https://cryptokittydex.com/pixels/0x895602d4a25404ca1e4404eb776c48ca10eaa6cb?page=%d'% (i),\
                 headers={'User-Agent': 'Mozilla/5.0'})
    webpage = urlopen(req).read()
    j=webpage.decode("utf-8")
    n=json.dumps(j)
    chick_lst.append(' '.join(re.findall('kitty-([\d+]+)',n)))

results=[]
results.append(" ".join(chick_lst))
chicken_lst=[int(i) for i in results[0].split()]

# Ship_lst

ship_lst = []

for i in range(1,29):
    time.sleep(4)
    req = Request('https://cryptokittydex.com/pixels/0xd9336b394c2c0c33f524a99cc45a1eaa63141df9?page=%d'% (i),\
                 headers={'User-Agent': 'Mozilla/5.0'})
    webpage = urlopen(req).read()
    j=webpage.decode("utf-8")
    n=json.dumps(j)
    ship_lst.append(' '.join(re.findall('kitty-([\d+]+)',n)))

results=[]
results.append(" ".join(ship_lst))
ships_lst=[int(i) for i in results[0].split()]

# Vamp_lst

vamp_lst = []

for i in range(1,17):
    time.sleep(4)
    req = Request('https://cryptokittydex.com/pixels/0x0aa95999e0e707e30806c554772550356c37711a?page=%d'% (i),\
                 headers={'User-Agent': 'Mozilla/5.0'})
    webpage = urlopen(req).read()
    j=webpage.decode("utf-8")
    n=json.dumps(j)
    vamp_lst.append(' '.join(re.findall('kitty-([\d+]+)',n)))

results=[]
results.append(" ".join(vamp_lst))
dracula_lst=[int(i) for i in results[0].split()]

# NY_lst

ny_lst = []

for i in range(1,13):
    time.sleep(4)
    req = Request('https://cryptokittydex.com/pixels/0xd9111e81e43181b2a47203cbb9e01cde70375e83?page=%d'% (i),\
                 headers={'User-Agent': 'Mozilla/5.0'})
    webpage = urlopen(req).read()
    j=webpage.decode("utf-8")
    n=json.dumps(j)
    ny_lst.append(' '.join(re.findall('kitty-([\d+]+)',n)))

results=[]
results.append(" ".join(ny_lst))
mist_lst=[int(i) for i in results[0].split()]

#New kitties list
nk_lst = [296750,296597,287880,296675,278621,275808,101,102,103,1]
          

fancy_lst = chicken_lst+ships_lst+dracula_lst+mist_lst+nk_lst

fancy=list(set(fancy_lst))

outfile = open('fancy14_12.txt', 'w')
for item in fancy:
    outfile.write("%s\n" % item)

with open("fancy14_12.txt") as file:
    lines = [line.strip() for line in file]

fancy = list(map(int, lines))

len(fancy)
