==> Cryptokitties/3_marking cattributes.ipynb <==
import pandas as pd
import numpy as np

columns = ['num','isGestating','isReady','cooldownIndex','nextActionAt',\
       'siringWithId','birthTime','matronId','sireId','generation','genes']
data = pd.read_csv('kitties_new.csv', header =None,index_col=None )
data.columns = columns

data['genes']=pd.to_numeric(data['genes'], errors='coerce')

## to_Hex

def to_hex(x):
    return hex(int(x))
data['genes'] = data['genes'].apply(lambda x: to_hex(x))

## cattributes

def breed(x):
    if x[61]=='5':
        return 'mainecoon'
    elif x[61]=='9':
        return 'cymric'
    elif x[61]=='6':
        return 'laperm'
    elif x[61]=='c':
        return 'munchkin'
    elif x[61]=='d':
        return 'sphynx'
    elif x[61]=='e':
        return 'ragamaffin'
    elif x[61]=='b':
        return 'himalayan'
    elif x[61]=='a':
        return 'chartreux'
    elif x[61]=='7':
        return 'persian'
    elif x[61]=='f':
        return 'ragdoll'
    else: return 'unknown'
data['breed']=data['genes'].apply(lambda x: breed(x))

def prim_col(x):
    if x[41]=='5':
        return 'mauveover'
    elif x[41]=='0': #mix or choose cloudwhite(12110 or 6.495%) or shadowgrey(23884 or 12.793%)
        return 'cloudgrey'
    elif x[41]=='1':
        return 'salmon'
    elif x[41]=='3':
        return 'orangesoda'
    elif x[41]=='6':
        return 'aquamarine'
    elif x[41]=='a':
        return 'greymatter'
    elif x[41]=='2':
        return 'oldlace'
    elif x[41]=='4':
        return 'cottoncandy'
    else: return 'unknown'
    
data['prim_col']=data['genes'].apply(lambda x: prim_col(x))

def sec_col(x):
    if x[31]=='2':
        return 'peach'
    elif x[31]=='3': 
        return 'bloodred'
    elif x[31]=='7':
        return 'emeraldgreen'
    elif x[31]=='4':
        return 'granitegrey'
    elif x[31]=='6':
        return 'kittencream'
    elif x[31]=='0':
        return 'belleblue'
    else: return 'unknown'

data['sec_col']=data['genes'].apply(lambda x: sec_col(x))       

def pat_col(x):
    if x[36]=='5':
        return 'cerulian'
    elif x[36]=='a': 
        return 'scarlet'
    elif x[36]=='6': #mix or skyblue(8662 or 4.631% ) or royalpurple(26412 or 14.113%)
        return 'blurple'
    elif x[36]=='b':
        return 'barkbrown'
    elif x[36]=='c':
        return 'coffee'
    elif x[36]=='d':
        return 'lemonade'
    elif x[36]=='8':
        return 'swampgreen'
    elif x[36]=='e':
        return 'chocolate'
    elif x[36]=='9':
        return 'violet'
    elif x[36]=='4':
        return 'wolfgrey'    
    else: return 'unknown'
data['pat_col']=data['genes'].apply(lambda x: pat_col(x)) 

def pat_type(x):
    if x[56]=='c':
        return 'spock'
    elif x[56]=='4': 
        return 'tigerpunk'
    elif x[56]=='8': 
        return 'calicool'
    elif x[56]=='9':
        return 'luckystripe'
    elif x[56]=='b':
        return 'jaguar'
    elif x[56]=='e':
        return 'totesbasic'
    elif x[56]=='f':
        return 'totesbasic'
    elif x[56]=='7':
        return 'totesbasic'
    else: return 'unknown'
data['pat_type']=data['genes'].apply(lambda x: pat_type(x)) 

def eye_type(x):
    if x[46]=='9':
        return 'wingtips'
    elif x[46]=='2': 
        return 'fabulous'
    elif x[46]=='4': 
        return 'otaku'
    elif x[46]=='3': #raisedbrow (32905 or 17.560%) or googly (3317 kitties or 1.770%)
        return 'raisedbrow'
    elif x[46]=='5':
        return 'simple'
    elif x[46]=='6':
        return 'crazy'
    elif x[46]=='7':
        return 'thicccbrowz'
    else: return 'unknown'
data['eye_type']=data['genes'].apply(lambda x: eye_type(x)) 

def eye_col(x):
    if x[51]=='1': #mix or gold (10,578 or 5.642%) or limegreen (10,250 or 5.466%)
        return 'golime'
    elif x[51]=='3'and x[28]=='1': 
        return 'mintgreen'
    elif x[51]=='3': 
        return 'bubblegum'
    elif x[51]=='6': 
        return 'chestnut'
    elif x[51]=='2':
        return 'topaz'
    elif x[51]=='7':
        return 'strawberry'
    elif x[51]=='5':
        return 'sizzurp'
    else: return 'unknown'
data['eye_col']=data['genes'].apply(lambda x: eye_col(x)) 

def mouth(x):
    if x[21]=='0': 
        return 'whixtensions'
    elif x[21]=='4': 
        return 'dali'
    elif x[21]=='a': 
        return 'saycheese'
    elif x[21]=='8': 
        return 'beard'
    elif x[21]=='7':
        return 'tongue'
    elif x[21]=='e':
        return 'happygokitty'
    elif x[21]=='9':
        return 'pouty'
    elif x[21]=='f':
        return 'soserious'
    elif x[21]=='3':
        return 'gerbil'
    else: return 'unknown'
data['mouth']=data['genes'].apply(lambda x: mouth(x)) 

data.to_csv('markedcats_13_12.csv', index=False)

#siringwith - партнер
#sireid - отец котенка
#matronid - мать котенка - 0 для гена 0
#siringWithId идентификатор беременности - исп для гена нового котенка
#nextActionAt - время спада колдауна - не актуально, так как не знаем какое было на момент продажи

df = pd.read_csv('markedcats_13_12.csv')
