==> Cryptokitties/4_kitties prices-working_ver.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
from matplotlib import pyplot as plt
from sklearn.preprocessing import LabelEncoder, OneHotEncoder
from sklearn.model_selection import train_test_split, cross_val_score

import datetime

%matplotlib inline

# Get data

cat_data = pd.read_csv('markedcats_13_12.csv')
cat_price=pd.read_csv('clear_price190k.csv')

with open("fancy14_12.txt") as file:
    lines = [line.strip() for line in file]
fancy = list(map(int, lines))

cat_data.head()

cat_price.shape

#remove unformatted data
cat_data=cat_data.drop(cat_data.index[0])

cat_data.shape

#### Mark or remove fancy

# check for fancy in later cells
cat_data['fancy']=cat_data['num'].apply(lambda x: 1 if x in fancy else 0)

cat_data[cat_data['fancy']==1].shape

#cat_data = cat_data[~cat_data['num'].isin(fancy)]

cat_data.shape

## Del or format features

#### Del gen 0

cat_data = cat_data[cat_data.generation != 0]
#cat_data = cat_data[cat_data.generation != 1]
#cat_data = cat_data[cat_data.generation != 2]

cat_data.shape

#### Del unnecessary features

cat_data=cat_data.drop('nextActionAt',axis=1)
cat_data=cat_data.drop('genes',axis=1)
#cat_data=cat_data.drop('siringWithId',axis=1)
cat_data=cat_data.drop('matronId',axis=1) #будут присутствовать у всех generation>0
cat_data=cat_data.drop('sireId',axis=1) #будут присутствовать у всех generation>0
#cat_data=cat_data.drop('birthTime',axis=1) #возможно стоит перевести в datetime и оставить день-месяц


cat_data.shape

#### Format true/false

cat_data.isGestating = cat_data.isGestating.astype(int)
cat_data.isReady = cat_data.isReady.astype(int)

# siringwith 
cat_data['siringWithId']=cat_data['siringWithId'].apply(lambda x: 1 if x!=0 else 0)

cat_data['birthTime'] = cat_data['birthTime'].apply(datetime.datetime.fromtimestamp)

## Ohe

#### label ohe

#check fancy
cat_df=pd.concat([cat_data[['num', 'isGestating', 'isReady','siringWithId','cooldownIndex', 'generation','fancy','birthTime']], pd.get_dummies(cat_data[['breed',
       'prim_col', 'sec_col', 'pat_col', 'pat_type', 'eye_type', 'mouth',
       'eye_col']])], axis=1)

#### ohe

temp=pd.get_dummies(cat_data['cooldownIndex'],prefix='cd')
cat_df=pd.concat([cat_df,temp],axis=1)
cat_df=cat_df.drop('cooldownIndex',axis=1)

temp2=pd.get_dummies(cat_data['generation'],prefix='gen')
cat_df=pd.concat([cat_df,temp2],axis=1)
cat_df=cat_df.drop('generation',axis=1)

## FEATURES

### rare

cat_df['dateBirth']=cat_df['birthTime'].apply(lambda x : x.date())

rare_feats=['eye_type_wingtips','pat_col_wolfgrey','breed_mainecoon','prim_col_oldlace',\
            'pat_col_cerulian','pat_col_violet','eye_type_fabulous','breed_chartreux',\
            'pat_type_jaguar','mouth_gerbil','prim_col_cottoncandy','mouth_dali','mouth_whixtensions',\
            'eye_col_bubblegum','sec_col_peach','eye_type_otaku','pat_type_tigerpunk','mouth_beard']

df_cs=pd.DataFrame(columns = rare_feats)
for i in rare_feats:
    df_c=cat_df.groupby(cat_df.dateBirth)[[i]].sum()
    df_cs[i]= df_c[i].cumsum()/cat_df.shape[0]
df_cs.columns=[s + '_cs' for s in rare_feats]
df_cs['dateBirth']=df_cs.index

merged_rare = pd.merge(left=cat_df,right=df_cs, how='left', left_on='dateBirth', right_on='dateBirth')

### common

common_feats=['pat_type_totesbasic', 'sec_col_granitegrey', 'sec_col_kittencream', \
              'pat_type_luckystripe', 'eye_type_thicccbrowz', 'eye_type_crazy', 'mouth_pouty',\
             'breed_himalayan', 'eye_col_sizzurp', 'breed_ragamaffin', 'mouth_happygokitty', 'mouth_soserious',\
              'eye_col_strawberry','prim_col_orangesoda','prim_col_aquamarine','mouth_saycheese','breed_munchkin',\
              'eye_type_raisedbrow','eye_type_simple','eye_col_topaz','breed_sphynx','pat_col_chocolate']

df_cs2=pd.DataFrame(columns = common_feats)
for i in common_feats:
    df_c2=cat_df.groupby(cat_df.dateBirth)[[i]].sum()
    df_cs2[i]= df_c2[i].cumsum()/cat_df.shape[0]
df_cs2.columns=[s + '_cs' for s in common_feats]
df_cs2['dateBirth']=df_cs2.index

merged_rare_common = pd.merge(left=merged_rare,right=df_cs2, how='left', left_on='dateBirth', right_on='dateBirth')

merged_rare_common.shape

cat_df=merged_rare_common

## Merge kitties features with kitties prices

cat_price.shape

#cat_price= cat_price.drop('Unnamed: 0', axis=1)

df= pd.merge(cat_df, cat_price, left_on = 'num', right_on = 'cat_id')
df.shape

## New date features

df['dateSold'] = df['timeStamp'].apply(datetime.datetime.fromtimestamp)

df['hourSold']=df['dateSold'].apply(lambda x: x.hour)

#list(set(df.hourSold))

df['weekday']=df['dateSold'].apply(lambda x: x.isoweekday())

df['weekend']=df['weekday'].apply(lambda x: 1 if x in [6,7] else 0)

df['evening']=df['hourSold'].apply(lambda x: 0 if x in range(3,18) else 1)

df['age']=df['dateSold'].apply(lambda x : x.date())-df['birthTime'].apply(lambda x : x.date())

df['age']=df['age']/np.timedelta64(1, 'D')
df['age']=df['age'].apply(lambda x: int(x))

#### ohe

temp=pd.get_dummies(df['hourSold'],prefix='h')
df=pd.concat([df,temp],axis=1)
df=df.drop('hourSold',axis=1)
temp2=pd.get_dummies(df['weekday'],prefix='wd')
df=pd.concat([df,temp2],axis=1)
df=df.drop('weekday',axis=1)

### Median price by day

df['date']=df['dateSold'].apply(lambda x : x.date())
df_medSold=df.groupby(df.dateSold.dt.date)[['soldPrice']].median()
df_medSold['date'] = df_medSold.index
df_medSold=df_medSold.rename(columns = {'soldPrice':'medPrice'})

### Mean price by day

df_meanSold=df.groupby(df.dateSold.dt.date)[['soldPrice']].mean()
df_meanSold['date'] = df_meanSold.index
df_meanSold=df_meanSold.rename(columns = {'soldPrice':'meanPrice'})

df_medSold

## Merge medianPrice, meanPrice with data

merged_left = pd.merge(left=df,right=df_medSold, how='left', left_on='date', right_on='date')
merged = pd.merge(left=merged_left,right=df_meanSold, how='left', left_on='date', right_on='date')

#merged['medDif']=merged['soldPrice']-merged['medPrice']

df=merged
df.shape

df = df.drop('num',axis=1)
df = df.drop('cat_id',axis=1)
df = df.drop('timeStamp',axis=1)
df = df.drop('dateSold',axis=1)

#df = df.drop('medPrice',axis=1)
df = df.drop('meanPrice',axis=1)
df = df.drop('date',axis=1)
df = df.drop('dateBirth',axis=1)

#df = df.drop('medDif',axis=1)
df = df.drop('birthTime',axis=1)

df['age']=np.log1p(df['age']+1)
df['medPrice']=np.log1p(df['medPrice'])

# OUTLIERS

%%time
out_train = df[['soldPrice']]
outliers_lst = list()
# Для каждого признака найдем слишком высокое или низкое значение
for feature in out_train.keys():
    
    # TODO: Считаем Q1 (25 процентов) 
    Q1 = np.percentile(out_train[feature],15)
    
    # TODO: Считаем Q1 (75 процентов)
    Q3 = np.percentile(out_train[feature],85)
    
    # TODO: Интерквартильный размах
    step = 1.5*(Q3-Q1)
    
    # Отображаем выбросы
    #print ("Возможные выбросы у признака '{}':".format(feature))
    #display(out_train[~((out_train[feature] >= Q1 - step) & (out_train[feature] <= Q3 + step))])
    
    for i in out_train[~((out_train[feature] >= Q1 - step) & (out_train[feature] <= Q3 + step))].index:
        outliers_lst.append(i)
    print(Q1,Q3)



%%time
# OPTIONAL: Выбираем выбросы от которых хотим избавиться
outliers  = []


outliers_dct = {x:outliers_lst.count(x) for x in outliers_lst}
for i in outliers_dct.keys():
    if outliers_dct[i]>0:
        outliers.append(i)
#print (outliers)
print (len(outliers))

good_data = df.drop(df.index[outliers]).reset_index(drop = True)

good_data.shape

### rename

df= good_data

# Model

y_train = np.log1p(df.soldPrice) #skrewed price distribution

X_train = df.drop('soldPrice', axis =1)

from sklearn.linear_model import Ridge

train_X, valid_X, train_y, valid_y = train_test_split(X_train, y_train, test_size = 0.1, random_state = 42)
model = Ridge(alpha=0.5, copy_X=True, fit_intercept=True, max_iter=None,
   normalize=False, random_state=42, solver='sag', tol=0.001)
model.fit(train_X,train_y)

model = GridSearchCV(Ridge(alpha=0.5, copy_X=True, fit_intercept=True, max_iter=None,
   normalize=False, random_state=42, solver='sag', tol=0.001), cv=3,
                   param_grid={"alpha": [0.3, 0.4, 0.5]})

model.fit(train_X,train_y)

pred = model.predict(valid_X)

valid = np.expm1(valid_y)
pred= np.expm1(pred)

from sklearn.metrics import mean_squared_error
from math import sqrt
rms = sqrt(mean_squared_error(valid, pred))

rms 

for x,y in zip(valid,pred):
    print(x,y)

from sklearn.feature_selection import RFE
from sklearn.linear_model import LinearRegression
 
lr = LinearRegression()
#select 5 the most informative features
rfe = RFE(lr, 20) 
selector = rfe.fit(train_X,train_y)

names = list(train_X.columns)
print (sorted(zip(map(lambda x: round(x, 4), selector.ranking_), names)))


from sklearn.linear_model import RandomizedLasso

Y = np.log1p(df.soldPrice) 
X = df.drop('soldPrice', axis =1)
names = [X.columns]
 
rlasso = RandomizedLasso(alpha=0.025)
rlasso.fit(X, Y)
 
print ("Features sorted by their score:")
print (sorted(zip(map(lambda x: round(x, 4), rlasso.scores_), 
                 names), reverse=True)
)

"""from sklearn.linear_model import (LinearRegression, Ridge, 
                                  Lasso, RandomizedLasso)
from sklearn.feature_selection import RFE, f_regression
from sklearn.preprocessing import MinMaxScaler
from sklearn.ensemble import RandomForestRegressor
import numpy as np
from minepy import MINE
 
np.random.seed(42)

Y = np.log1p(df.soldPrice) #skrewed price distribution
X = df.drop('soldPrice', axis =1)
 
names = [X.columns]
 
ranks = {}
 
def rank_to_dict(ranks, names, order=1):
    minmax = MinMaxScaler()
    ranks = minmax.fit_transform(order*np.array([ranks]).T).T[0]
    ranks = map(lambda x: round(x, 2), ranks)
    return dict(zip(names, ranks ))
 
lr = LinearRegression(normalize=True)
lr.fit(X, Y)
ranks["Linear reg"] = dict(zip(np.abs(lr.coef_), names)
 
ridge = Ridge(alpha=7)
ridge.fit(X, Y)
ranks["Ridge"] = rank_to_dict(np.abs(ridge.coef_), names)
 
lasso = Lasso(alpha=.05)
lasso.fit(X, Y)
ranks["Lasso"] = rank_to_dict(np.abs(lasso.coef_), names)
 
rlasso = RandomizedLasso(alpha=0.04)
rlasso.fit(X, Y)
ranks["Stability"] = rank_to_dict(np.abs(rlasso.scores_), names)
 
#stop the search when 5 features are left (they will get equal scores)
rfe = RFE(lr, n_features_to_select=5)
rfe.fit(X,Y)
ranks["RFE"] = rank_to_dict(map(float, rfe.ranking_), names, order=-1)
 
rf = RandomForestRegressor()
rf.fit(X,Y)
ranks["RF"] = rank_to_dict(rf.feature_importances_, names)
 
f, pval  = f_regression(X, Y, center=True)
ranks["Corr."] = rank_to_dict(f, names)
 
mine = MINE()
mic_scores = []
for i in range(X.shape[1]):
    mine.compute_score(X[:,i], Y)
    m = mine.mic()
    mic_scores.append(m)
    
ranks["MIC"] = rank_to_dict(mic_scores, names) 
 
r = {}
for name in names:
    r[name] = round(np.mean([ranks[method][name] 
                             for method in ranks.keys()]), 2)

methods = sorted(ranks.keys())
ranks["Mean"] = r
methods.append("Mean")
 
print ("\t%s" % "\t".join(methods))
for name in names:
    print ("%s\t%s" % (name, "\t".join(map(str, 
                         [ranks[method][name] for method in methods]))))
""";

# Model for last days

df=merged

df.shape

df.head()
df = df.sort_values('date')

df.tail()

df = df.drop('num',axis=1)
df = df.drop('cat_id',axis=1)
df = df.drop('timeStamp',axis=1)
df = df.drop('dateSold',axis=1)

#df = df.drop('medPrice',axis=1)
df = df.drop('meanPrice',axis=1)
df = df.drop('date',axis=1)
df = df.drop('dateBirth',axis=1)

#df = df.drop('medDif',axis=1)
df = df.drop('birthTime',axis=1)

df['age']=np.log1p(df['age']+1)
df['medPrice']=np.log1p(df['medPrice'])

#total 124820
df_last = df[100000:]


df_last.shape

y_train = np.log1p(df_last.soldPrice) #skrewed price distribution

X_train = df_last.drop('soldPrice', axis =1)

from sklearn.linear_model import Ridge

train_X, valid_X, train_y, valid_y = train_test_split(X_train, y_train, test_size = 0.1, random_state = 42)
model = Ridge(alpha=0.5, copy_X=True, fit_intercept=True, max_iter=None,
   normalize=False, random_state=42, solver='sag', tol=0.001)
model.fit(train_X,train_y)

pred = model.predict(valid_X)

valid = np.expm1(valid_y)
pred= np.expm1(pred)

rms = sqrt(mean_squared_error(valid, pred))

rms 

for x,y in zip(valid,pred):
    print(x,y)

# CV

from sklearn.model_selection import cross_val_score
from sklearn import metrics

clf = Ridge(alpha=0.5, copy_X=True, fit_intercept=True, max_iter=None,
   normalize=False, random_state=42, solver='sag', tol=0.001)
train_X, valid_X, train_y, valid_y = train_test_split(X_train, y_train, test_size = 0.1, random_state = 42)

scores = cross_val_score(clf, train_X,train_y, cv=5, scoring='neg_mean_squared_error')


from sklearn.model_selection import GridSearchCV
model = GridSearchCV(Ridge(alpha=0.5, copy_X=True, fit_intercept=True, max_iter=None,
   normalize=False, random_state=42, solver='sag', tol=0.001), cv=3,
                   param_grid={"alpha": [0.3, 0.4, 0.5]})
model.fit(train_X,train_y)

model.best_estimator_

pred = model.predict(valid_X)
valid = np.expm1(valid_y)
pred= np.expm1(pred)
valid = np.expm1(valid_y)
pred= np.expm1(pred)
res =sqrt(mean_squared_error(valid, pred))
print (res)
